Move the automatic-transformation loop body into a Zany mixin
PhysicallyMappedElement (physically_mapped.py) is restored to its
original shape: a generic, abstract mixin with no knowledge of the
zany theory, exactly as used by hand-coded elements (AW, HCT,
PowellSabin, Walkington, ...). It no longer imports finat.functional.

The template-method loop and its four hooks move into a new
ZanyPhysicallyMappedElement(PhysicallyMappedElement) in finat/zany.py,
which ScalarPhysicallyMappedElement and PiolaPhysicallyMappedElement
now subclass instead of PhysicallyMappedElement directly. This keeps
all zany-specific machinery -- the loop, the hooks, the pure math
functions, and the symbolic PhysicallyMappedFunctional dependency --
contained in finat/zany.py and finat/functional.py, with
physically_mapped.py staying general-purpose infrastructure.

No behavioral change: full finat suite, flake8, and pydocstyle pass
unchanged.
