+++ finat/zany.py
@@ -2,17 +2,17 @@ r"""Automatic basis transformations for physically mapped elements.
 
 This module automates the transformation theory of Kirby (2017) and
 Brubeck & Kirby (2025).  Given a FIAT element whose degrees of freedom
-are not preserved under push-forward, :class:`~finat.physically_mapped.
-PhysicallyMappedElement` constructs the matrix :math:`V` relating the
-reference nodes to the push-forwards of the physical nodes, so that the
-physical basis functions are obtained as :math:`M F^*(\hat\Psi)` with
-:math:`M = V^T`.
+are not preserved under push-forward, :class:`ZanyPhysicallyMappedElement`
+constructs the matrix :math:`V` relating the reference nodes to the
+push-forwards of the physical nodes, so that the physical basis
+functions are obtained as :math:`M F^*(\hat\Psi)` with :math:`M = V^T`.
 
 Degrees of freedom are represented symbolically by
 :class:`finat.functional.PhysicallyMappedFunctional` and processed
-generically, without dispatching over FIAT functional types.  This
-module supplies the two mapping-specific mixins that plug into the
-entity-by-entity assembly loop:
+generically, without dispatching over FIAT functional types.
+:class:`ZanyPhysicallyMappedElement` implements the entity-by-entity
+assembly loop once, calling four hooks that carry all mapping-specific
+knowledge; this module supplies the two mixins implementing them:
 
 * :class:`ScalarPhysicallyMappedElement`, for scalar elements with an
   affine (identity) pullback -- Morley, Hermite, Argyris, Bell.  Each
@@ -42,8 +42,9 @@ from operator import add
 import numpy
 
 from FIAT.finite_element import FiniteElement
-from gem import Literal, Node, Power, Zero
-from finat.physically_mapped import PhysicallyMappedElement, adjugate, determinant
+from gem import Literal, ListTensor, Node, Power, Zero
+from finat.functional import PhysicallyMappedFunctional
+from finat.physically_mapped import PhysicallyMappedElement, adjugate, determinant, identity
 
 
 def generalized_cross(tangents) -> numpy.ndarray:
@@ -139,7 +140,146 @@ def _weight_ratio(wi: numpy.ndarray, wj: numpy.ndarray, tol: float) -> float:
     return s
 
 
-class ScalarPhysicallyMappedElement(PhysicallyMappedElement):
+class ZanyPhysicallyMappedElement(PhysicallyMappedElement):
+    r"""Mixin implementing the entity-by-entity assembly loop shared by
+    :class:`ScalarPhysicallyMappedElement` and
+    :class:`PiolaPhysicallyMappedElement`.
+
+    Following the factorization :math:`V = E V^c D` of Kirby (2017) and
+    Brubeck & Kirby (2025), the matrix :math:`V` relating the reference
+    nodes to the push-forwards of the physical nodes is assembled one
+    topological entity at a time, in increasing dimension, so that the
+    completion of a node on an entity can always be resolved against
+    the already-assembled rows of lower-dimensional entities.
+
+    On each entity, nodes that are already push-forward invariant
+    (:meth:`_invariant_dofs`) contribute an identity row for free,
+    since :math:`V` starts out as the identity.  The rest are assembled
+    by :meth:`_facet_dof_rows` (on a codimension-1 entity) or
+    :meth:`_point_dof_rows` (elsewhere); these hooks, together with
+    :meth:`_check_mapping`, encode the mapping-specific (affine or
+    Piola) part of the theory, and :meth:`basis_transformation` itself
+    contains no knowledge of which mapping is in play.
+    """
+
+    #: Numerical tolerance used throughout automatic basis transformation
+    #: to detect vanishing coefficients.
+    tol = 1e-12
+
+    def basis_transformation(self, coordinate_mapping) -> ListTensor:
+        fiat_element = self._element
+        self._check_mapping(fiat_element)
+
+        ref_el = fiat_element.get_reference_element()
+        sd = ref_el.get_spatial_dimension()
+        bary, = ref_el.make_points(sd, 0, sd + 1)
+        J = coordinate_mapping.jacobian_at(bary)
+
+        nodes = fiat_element.dual_basis()
+        V = identity(fiat_element.space_dimension())
+
+        processed = set()
+        entity_ids = fiat_element.entity_dofs()
+        for dim in sorted(entity_ids):
+            for entity in sorted(entity_ids[dim]):
+                group = {i: PhysicallyMappedFunctional.from_fiat(nodes[i])
+                         for i in entity_ids[dim][entity]}
+                invariant = self._invariant_dofs(group, dim, sd)
+                processed.update(invariant)
+                group = {i: ell for i, ell in group.items() if i not in invariant}
+                if not group:
+                    continue
+                if dim == sd - 1:
+                    self._facet_dof_rows(V, group, fiat_element, entity, J, processed)
+                else:
+                    self._point_dof_rows(V, group, fiat_element, J, processed)
+
+        _rescale_derivative_dofs(V, fiat_element, coordinate_mapping)
+        ndof = self.space_dimension()
+        return ListTensor(V[:, :ndof].T)
+
+    def _check_mapping(self, fiat_element):
+        """Verify that this class knows how to transform this element's pullback.
+
+        :arg fiat_element: The FIAT element defined on the reference cell.
+        :raises NotImplementedError: If the pullback is not supported.
+        """
+        raise NotImplementedError(
+            f"{type(self).__name__} does not implement automatic basis transformation.")
+
+    def _invariant_dofs(self, group, dim, sd):
+        """Select the nodes of an entity that are already push-forward invariant.
+
+        :arg group: Dict mapping node index to :class:`PhysicallyMappedFunctional`
+            for the reference nodes associated with one entity.
+        :arg dim: Topological dimension of the entity.
+        :arg sd: Spatial dimension of the cell.
+        :returns: The subset of ``group`` keys whose row of :math:`V` is
+            the identity row.
+        """
+        raise NotImplementedError(
+            f"{type(self).__name__} does not implement automatic basis transformation.")
+
+    def _facet_dof_rows(self, V, group, fiat_element, entity, J, processed):
+        """Assemble the rows of V for the non-invariant nodes on a facet.
+
+        :arg V: Object array being assembled; rows are set in place.
+        :arg group: Dict mapping node index to :class:`PhysicallyMappedFunctional`
+            for the non-invariant reference nodes on this facet.
+        :arg fiat_element: The FIAT element defined on the reference cell.
+        :arg entity: The facet number.
+        :arg J: GEM expression for the cell Jacobian.
+        :arg processed: Indices of the already assembled rows of ``V``;
+            updated in place.
+        """
+        raise NotImplementedError(
+            f"{type(self).__name__} does not implement automatic basis transformation.")
+
+    def _point_dof_rows(self, V, group, fiat_element, J, processed):
+        """Assemble the rows of V for the non-invariant nodes away from a facet.
+
+        :arg V: Object array being assembled; rows are set in place.
+        :arg group: Dict mapping node index to :class:`PhysicallyMappedFunctional`
+            for the non-invariant reference nodes on this entity.
+        :arg fiat_element: The FIAT element defined on the reference cell.
+        :arg J: GEM expression for the cell Jacobian.
+        :arg processed: Indices of the already assembled rows of ``V``;
+            updated in place.
+        """
+        raise NotImplementedError(
+            f"{type(self).__name__} does not implement automatic basis transformation.")
+
+
+def _rescale_derivative_dofs(V, fiat_element, coordinate_mapping):
+    r"""Rescale derivative degrees of freedom by the cell size.
+
+    Each physical node of derivative order :math:`m` is redefined with a
+    factor :math:`h^{-m}`, where :math:`h` averages the cell size over
+    the vertices of its entity.  This is the FInAT convention keeping
+    the mass matrix well-conditioned; it is consistent across cells
+    because the scaling only depends on shared entities.
+
+    :arg V: Object array being assembled; columns are rescaled in place.
+    :arg fiat_element: The FIAT element defined on the reference cell.
+    :arg coordinate_mapping: Object providing the physical geometry as
+        GEM expressions.
+    """
+    # cell_size may be a GEM expression or a numpy array of numbers
+    h = coordinate_mapping.cell_size()
+    top = fiat_element.get_reference_element().get_topology()
+    nodes = fiat_element.dual_basis()
+    entity_ids = fiat_element.entity_dofs()
+    for dim in entity_ids:
+        for entity in entity_ids[dim]:
+            verts = top[dim][entity]
+            havg = reduce(add, (h[v] for v in verts)) / len(verts)
+            for i in entity_ids[dim][entity]:
+                order = nodes[i].max_deriv_order
+                if order > 0:
+                    V[:, i] = V[:, i] * havg**(-order)
+
+
+class ScalarPhysicallyMappedElement(ZanyPhysicallyMappedElement):
     r"""Mixin deriving the basis transformation for a scalar element
     with an affine (identity) pullback.
 
@@ -270,7 +410,7 @@ def _scalar_point_rows(V: numpy.ndarray, group: dict, J: Node,
             processed.add(i)
 
 
-class PiolaPhysicallyMappedElement(PhysicallyMappedElement):
+class PiolaPhysicallyMappedElement(ZanyPhysicallyMappedElement):
     r"""Mixin deriving the basis transformation for a vector- or
     tensor-valued element under the (double) contravariant Piola
     pullback.
